Turns Location into a class. Makes amounts and exchange rates BigDecimals due to float rounding issues. OrderService.list methods return lists instead of collections, for ordering.

+++ src/main/java/com/google/gfx/services/exchangerate/ExchangeRateService.java
@@ -1,5 +1,6 @@
 package com.google.gfx.services.exchangerate;
 
+import java.math.BigDecimal;
 import java.util.Currency;
 
 /**
@@ -19,5 +20,5 @@ public interface ExchangeRateService {
    * @param toCurrency gotten currency
    * @return the exchange rate value, such that 'itself * fromCurrency = toCurrency'
    */
-  float getExchangeRate(Currency fromCurrency, Currency toCurrency);
+  BigDecimal getExchangeRate(Currency fromCurrency, Currency toCurrency);
 }
